Ensure hosts file either gets left alone or a new one created
Previously, if the number of lines in the unsorted temp hosts file was less than the threshold *and* an existing hosts file didn't exist, then a new hosts file would not be created. This commit fixes that situation, to ensure that — at the very least — a less-than-adequately-populated hosts file exists.